MDEV-21791: Packed Sort Keys: provide details in ANALYZE FORMAT=JSON.
It's useful to know whether packed addon fields or sort-keys are used
by the query.

Unfortunately it's hard show such information in EXPLAIN[FORMAT=JSON]
(and even if it was possible, it's not clear whether this level of
detail would be appropriate there).

Provide this information in ANALYZE FORMAT=JSON output.